update clean architecture
A change is made in the update of the files to be able to use the created folders (start and middlewares), in order to follow the principles of capable in clean architecture.
The reason for creating a home folder for the indexes of the different functionalities is to comply with the principle of singularity (unique functions in each created file).